refactor: 홈, 어드민 페이지 분리 및 홈 페이지 내부 UI 분리, formatPrice 유틸리티 개선
