+++ docs/refactoring-process.md
@@ -704,18 +704,76 @@ import Icon from '@/basic/components/icons/Icon';
   - **성능**: 최적화된 이벤트 처리로 성능 향상
   - **확장성**: 새로운 기능 추가가 용이한 구조
 
+### 2.12 App.tsx 대폭 리팩토링 ✅
+
+- **목적**: 거대한 단일 컴포넌트를 간결한 라우팅 컴포넌트로 변환하여 관심사 분리
+- **작업 내용**:
+  - [x] **App.tsx 대폭 간소화**:
+    - 95줄에서 19줄로 대폭 감소 (767줄 삭제)
+    - 거대한 단일 컴포넌트를 간결한 라우팅 컴포넌트로 변환
+    - 단순한 상태 관리와 페이지 전환 로직만 유지
+  - [x] **페이지 컴포넌트 분리**:
+    - `HomePage.tsx`: 쇼핑몰 홈 페이지 로직 분리 (482줄 삭제)
+    - `AdminPage.tsx`: 관리자 페이지 로직 분리 (95줄 추가)
+    - `pages/index.ts`: 페이지 컴포넌트 통합 export
+  - [x] **레이아웃 컴포넌트 추가**:
+    - `DashBoardLayout.tsx`: 관리자 대시보드 레이아웃
+  - [x] **테스트 안정성 확보**:
+    - `addNotification`을 AdminPage에 전달하여 테스트 실패 해결
+    - 알림 시스템 연결 문제 해결
+  - [x] **기타 컴포넌트 수정**:
+    - Cart, Product, Coupon 관련 컴포넌트들의 import 경로 수정
+    - useCart 훅의 import 경로 수정
+- **검증**:
+  - [x] TypeScript 컴파일 오류 없음
+  - [x] 모든 기능 정상 작동
+  - [x] 컴포넌트 간 결합도 감소
+  - [x] 테스트 안정성 확보
+- **개선 효과**:
+  - **코드 구조 개선**: 단일 책임 원칙(SRP) 적용
+  - **관심사 분리**: 각 페이지별 독립적 관리 가능
+  - **유지보수성 향상**: 변경 시 영향 범위 최소화
+  - **확장성 확보**: 새로운 페이지 추가 용이
+  - **테스트 용이성**: 각 페이지를 독립적으로 테스트 가능
+
+### 2.13 formatPrice 함수 수정 ✅
+
+- **목적**: 테스트에서 요구하는 가격 표시 형식에 맞게 함수 수정
+- **작업 내용**:
+  - [x] **문제 분석**:
+    - 테스트에서 찾는 텍스트: `"25,000원"`
+    - 실제 표시되는 텍스트: `"25,000 대한민국 원"`
+    - formatPrice 함수의 잘못된 옵션 사용
+  - [x] **formatPrice 함수 수정**:
+    - `formatPrice.unit()`: `"5,000원"` 형식 반환
+    - `formatPrice.currency()`: `"₩5,000"` 형식 반환
+    - Intl.NumberFormat API 올바르게 활용
+  - [x] **테스트 호환성 확보**:
+    - 테스트에서 요구하는 형식에 맞게 수정
+    - 가격 표시 형식 통일
+- **검증**:
+  - [x] TypeScript 컴파일 오류 없음
+  - [x] 가격 표시 형식 정상 작동
+  - [x] 테스트 호환성 확보
+- **개선 효과**:
+  - **테스트 안정성**: 가격 표시 형식 통일로 테스트 통과
+  - **일관성**: 모든 가격 표시가 동일한 형식 사용
+  - **유지보수성**: formatPrice 함수의 명확한 역할 분리
+
 ## 📊 현재 작업 진행 상황 업데이트
 
-| 작업                    | 상태      | 완성도 | 주요 고민사항                   |
-| ----------------------- | --------- | ------ | ------------------------------- |
-| Header 리팩토링         | ✅ 완료   | 100%   | Props Drilling vs 컴포지션 패턴 |
-| 컴포넌트 분리           | ✅ 완료   | 100%   | 컴포넌트 분리 수준 결정         |
-| 디바운싱 분석           | ✅ 완료   | 100%   | 디바운싱 vs 쓰로틀링            |
-| Icon 시스템             | 🔄 진행중 | 60%    | 복잡한 시스템 vs 간단한 시스템  |
-| App.tsx 업데이트        | ✅ 완료   | 100%   | 타입 안전성 vs 간단함           |
-| ProductList/ProductCard | ✅ 완료   | 100%   | 재사용성 vs 특화성              |
-| 레이아웃 시스템         | ✅ 완료   | 100%   | 일관성 vs 유연성                |
-| useLocalStorage 고도화  | ✅ 완료   | 100%   | 성능 vs 기능성                  |
+| 작업                    | 상태      | 완성도 | 주요 고민사항                       |
+| ----------------------- | --------- | ------ | ----------------------------------- |
+| Header 리팩토링         | ✅ 완료   | 100%   | Props Drilling vs 컴포지션 패턴     |
+| 컴포넌트 분리           | ✅ 완료   | 100%   | 컴포넌트 분리 수준 결정             |
+| 디바운싱 분석           | ✅ 완료   | 100%   | 디바운싱 vs 쓰로틀링                |
+| Icon 시스템             | 🔄 진행중 | 60%    | 복잡한 시스템 vs 간단한 시스템      |
+| App.tsx 업데이트        | ✅ 완료   | 100%   | 타입 안전성 vs 간단함               |
+| ProductList/ProductCard | ✅ 완료   | 100%   | 재사용성 vs 특화성                  |
+| 레이아웃 시스템         | ✅ 완료   | 100%   | 일관성 vs 유연성                    |
+| useLocalStorage 고도화  | ✅ 완료   | 100%   | 성능 vs 기능성                      |
+| App.tsx 대폭 리팩토링   | ✅ 완료   | 100%   | 거대 컴포넌트 분리 및 테스트 안정성 |
+| formatPrice 함수 수정   | ✅ 완료   | 100%   | 테스트 호환성 및 가격 표시 형식     |
 
 ## 🎯 핵심 성과 및 인사이트
 
@@ -750,12 +808,14 @@ import Icon from '@/basic/components/icons/Icon';
 1. ✅ **Phase 2.8: Header 컴포넌트 분리 및 리팩토링** - 컴포지션 패턴 적용 (완료)
 2. ✅ **Phase 2.10: 컴포넌트 분리 및 레이아웃 시스템 구축** - ProductList, ProductCard, 페이지 분리 (완료)
 3. ✅ **Phase 2.11: useLocalStorage 훅 고도화** - 탭 간 동기화 및 성능 최적화 (완료)
+4. ✅ **Phase 2.12: App.tsx 대폭 리팩토링** - 거대 컴포넌트 분리 및 페이지 컴포넌트 분리 (완료)
+5. ✅ **Phase 2.13: formatPrice 함수 수정** - 테스트 호환성 및 가격 표시 형식 개선 (완료)
 
 ### 기본과제 (남은 작업)
 
-4. **Phase 2.9: Icon 시스템 완성** - 네임스페이스 패턴 최적화 (다음 우선순위)
-5. **Phase 2.999: 긴 함수 분할** - 20줄 이하 함수 규칙 준수
-6. **Phase 3.1: 엔티티 컴포넌트와 UI 컴포넌트 분리** - README 요구사항
+6. **Phase 2.9: Icon 시스템 완성** - 네임스페이스 패턴 최적화 (다음 우선순위)
+7. **Phase 2.999: 긴 함수 분할** - 20줄 이하 함수 규칙 준수
+8. **Phase 3.1: 엔티티 컴포넌트와 UI 컴포넌트 분리** - README 요구사항
 
 ### 심화과제 (기본과제 완료 후)
 
@@ -764,4 +824,4 @@ import Icon from '@/basic/components/icons/Icon';
 
 ---
 
-**마지막 업데이트**: 2024년 (컴포넌트 분리 및 레이아웃 시스템 구축, useLocalStorage 훅 고도화, ProductList/ProductCard 분리 완료)
+**마지막 업데이트**: 2024년 (App.tsx 대폭 리팩토링, formatPrice 함수 수정, 페이지 컴포넌트 분리 완료)
