+++ src/basic/features/product/components/ProductList.tsx
@@ -1,3 +1,4 @@
+import { Coupon } from "@/basic/features/coupon/types/coupon.type";
 import { AddNotification } from "@/basic/features/notification/types/notification";
 import ProductCard from "@/basic/features/product/components/ProductCard";
 import { useProducts } from "@/basic/features/product/hooks/useProducts";
@@ -7,11 +8,15 @@ import { ProductWithUI } from "@/basic/features/product/types/product";
 interface ProductListProps {
   searchTerm: string;
   addNotification: AddNotification;
+  selectedCoupon: Coupon | null;
+  setSelectedCoupon: (coupon: Coupon | null) => void;
 }
 
 export default function ProductList({
   searchTerm,
   addNotification,
+  selectedCoupon,
+  setSelectedCoupon,
 }: ProductListProps) {
   const { products } = useProducts({
     addNotification,
@@ -42,6 +47,8 @@ export default function ProductList({
               key={product.id}
               product={product}
               addNotification={addNotification}
+              selectedCoupon={selectedCoupon}
+              setSelectedCoupon={setSelectedCoupon}
             />
           ))}
         </div>
