+++ src/basic/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as HomePage } from "./HomePage";
+export { default as AdminPage } from "./AdminPage";
